workshops/Medical_Imaging_AI/source/requirements.txt
monai-weekly[gdown, nibabel, tqdm, ignite]
matplotlib
psutil